付费用户组 删除

+++ app/Api/Controller/GroupV3/BatchDeleteGroupsController.php
@@ -20,6 +20,7 @@ namespace App\Api\Controller\GroupV3;
 
 use App\Common\ResponseCode;
 use App\Models\Group;
+use App\Models\GroupPaidUser;
 use App\Models\GroupUser;
 use App\Models\GroupUserMq;
 use App\Models\User;
@@ -27,6 +28,7 @@ use App\Repositories\UserRepository;
 use Discuz\Base\DzqAdminController;
 use Discuz\Base\DzqLog;
 use Illuminate\Contracts\Bus\Dispatcher;
+use Illuminate\Support\Carbon;
 
 class BatchDeleteGroupsController extends DzqAdminController
 {
@@ -67,9 +69,17 @@ class BatchDeleteGroupsController extends DzqAdminController
         $groupIdDeletes = $groupDatas->pluck('id')->toArray();
         $result = $this->changeUserGroupId($groupIdDeletes);
         if ($result == 'true'){
-            $groupDatas->each(function ($group) {
+            $paidGroupIds=[];
+            $groupDatas->each(function ($group) use(&$paidGroupIds) {
+                if ($group->is_paid == Group::IS_PAID){
+                    $paidGroupIds[] = $group->id;
+                }
                 $group->delete();
             });
+
+            GroupPaidUser::query()->whereIn('group_id', $paidGroupIds)
+                ->update(['operator_id' => $this->actor->id, 'deleted_at' => Carbon::now(), 'delete_type' => GroupPaidUser::DELETE_TYPE_ADMIN]);
+
             return $this->outPut(ResponseCode::SUCCESS, '');
         }else{
             return $this->outPut(ResponseCode::INTERNAL_ERROR, $result);
@@ -77,46 +87,46 @@ class BatchDeleteGroupsController extends DzqAdminController
     }
 
     private function changeUserGroupId($groupIdDeletes){
-        $do_userIds=[];
         try {
             $defualtGroup = Group::query()->select('id')->where("default",1)->first();
             $defualtGroupId = Group::MEMBER_ID;
             if (!empty($defualtGroup)){
-                $defualtGroupId = $defualtGroupId->id;
+                $defualtGroupId = $defualtGroup->id;
             }
             $changeUserIds = GroupUser::query()->select("user_id")->whereIn('group_id',$groupIdDeletes)->get();
 
-            $doClosure = function () use ($changeUserIds, $groupIdDeletes, $defualtGroupId){
+            $dbMgr = app('db');
+            $dbResult = $dbMgr->transaction(function () use ($changeUserIds, $groupIdDeletes, $defualtGroupId, $dbMgr){
                 foreach ($changeUserIds as $key=>$item){
                     $user_id = $item->user_id;
 
                     //减少过期时间
-                    $remainDays = GroupUserMq::query()->whereIn('group_id',$groupIdDeletes)->sum('remain_days');
-                    User::query()->where('id',$user_id)->update(['expired_at'=>'date_add(`expired_at`,interval '.-$remainDays.' day)']);
+                    $remainDays = GroupUserMq::query()->whereIn('group_id',$groupIdDeletes)->where('user_id',$user_id)->sum('remain_days');
+                    User::query()->where('id',$user_id)->update(['expired_at'=>$dbMgr->raw("DATE_ADD(expired_at,interval ".-$remainDays." day)")]);
 
                     //设置新的用户组
-                    $newFeeGroup = GroupUserMq::query()->select("tb1.group_id")
-                        ->leftJoin("groups as tb2","tb1.group_id","tb2.id")
+                    $newFeeGroup = GroupUserMq::query()->select("group_id")
+                        ->leftJoin("groups as tb2","group_id","tb2.id")
                         ->where("tb2.is_paid",Group::IS_PAID)
-                        ->where('tb1.user_id', $user_id)
-                        ->whereNotIn('tb1.group_id',$groupIdDeletes)
+                        ->where('user_id', $user_id)
+                        ->whereNotIn('group_id',$groupIdDeletes)
                         ->orderByDesc("tb2.level")->first();
                     if (empty($newFeeGroup)){
                         //设置为免费默认组
-                        GroupUser::query()->where('user_id',$user_id)->update('group_id',$defualtGroupId);
+                        GroupUser::query()->where('user_id',$user_id)->update(['group_id'=>$defualtGroupId]);
                     }else{
-                        GroupUser::query()->where('user_id',$user_id)->update('group_id',$newFeeGroup->group_id);
+                        GroupUser::query()->where('user_id',$user_id)->update(['group_id'=> $newFeeGroup->group_id]);
                     }
-                    $do_userIds[] = $user_id;
                 }
-            };
-            $doClosure();
+                return true;
+            });
+            if (!$dbResult){
+                return false;
+            }
 
             GroupUserMq::query()->whereIn('group_id',$groupIdDeletes)->delete();
             return true;
         }catch (Throwable $e){
-            GroupUserMq::query()->whereIn('group_id',$groupIdDeletes)->whereIn('user_id',$do_userIds)->delete();
-
             if (empty($e->validator) || empty($e->validator->errors())) {
                 $errorMsg = $e->getMessage();
             } else {
